fastAPi-Carte/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
motor>=3.3.2
pydantic>=2.5.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
python-jose>=3.3.0
passlib>=1.7.4